% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_streets.R
\name{get_streets}
\alias{get_streets}
\title{Get streets shape file}
\usage{
get_streets(
  filename = here::here("../data/STREETS 2/STREETS.shp"),
  layer = "STREETS"
)
}
\arguments{
\item{filename}{Streets shape file}

\item{layer}{what layer to pull from}
}
\value{
sf object
}
\description{
Expects a streets shapefile with line features
}
