% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidents_as_sf.R
\name{incidents_as_sf}
\alias{incidents_as_sf}
\title{Converts EMS incidents into sf object}
\usage{
incidents_as_sf(
  incidents,
  coords = c("scene_lon", "scene_lat"),
  crs = "EPSG:4326"
)
}
\arguments{
\item{incidents}{dataframe with incidents}

\item{coords}{pair-list of latitude and longitude columns}

\item{crs}{input string for crs, if one hasn't been determined when importing
streets using `get_streets()`.}
}
\value{
sf object
}
\description{
Takes incidents with latitude and longitude data and returns an sf object.
Notably only includes incidents with coordinates present.
}
