% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imports.R
\name{int_len}
\alias{int_len}
\title{Get the duration in seconds between two points in time}
\usage{
int_len(start, end)
}
\arguments{
\item{start}{Date times}

\item{end}{Date times}
}
\value{
duration in seconds
}
\description{
See the specifications for `lubridate::interval`
}
