% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_incidents_streets.R
\name{join_incidents_streets}
\alias{join_incidents_streets}
\title{Join EMS incidents with nearby streets}
\usage{
join_incidents_streets(sf_incidents, streets, distance, distance_units)
}
\arguments{
\item{sf_incidents}{sf object of EMS incidents}

\item{streets}{sf object of streets}

\item{distance}{numeric distance for points to be within}

\item{distance_units}{character string of units to pass to `units::set_units`}
}
\value{
sf object of join
}
\description{
Joins EMS incidents with the Streets list based on nearby streets, based on
a given distance.
}
