% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_mvcs_streets.R
\name{merge_mvcs_streets}
\alias{merge_mvcs_streets}
\title{Get Street Attribute data for MVC incidents}
\usage{
merge_mvcs_streets(sf_incidents, streets)
}
\arguments{
\item{sf_incidents}{MVCs, as sf}

\item{streets}{Streets sf object}
}
\value{
Returns dataframe with spatial data joined in.
}
\description{
Completes a spatial join to get street attribute data for
MVCs within a specified distance of any given road.
}
