% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_incidents.R
\name{plot_incidents}
\alias{plot_incidents}
\title{Plots incidents alone}
\usage{
plot_incidents(sf_incidents)
}
\arguments{
\item{sf_incidents}{sf object of incidents}
}
\value{
Outputs a ggplot2 plot of the incidents.
}
\description{
Plots incidents alone, using included latitude and longitude.
}
