% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_streets.R
\name{plot_streets}
\alias{plot_streets}
\title{Plots Streets}
\usage{
plot_streets(streets, classes)
}
\arguments{
\item{streets}{sf data frame of streets}

\item{classes}{character list of class types to include.}
}
\value{
ggplot element
}
\description{
Plots streets data using `sf` and `ggplot`, optionally filtering by
class. Expects streets to have a column named "CLASS" to filter against.
}
