% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_merge.R
\name{pmerge_mvcs_streets}
\alias{pmerge_mvcs_streets}
\title{Get Street Attribute data for MVC incidents Parallel}
\usage{
pmerge_mvcs_streets(
  mvcs,
  sf_incidents,
  streets,
  distance = units::set_units(10, "m")
)
}
\arguments{
\item{mvcs}{MVCs, as sf}

\item{sf_incidents}{MVCs, as sf}

\item{streets}{Streets sf object}

\item{distance}{Distance units}
}
\value{
Returns dataframe with spatial data joined in.
}
\description{
Completes a spatial join to get street attribute data for
MVCs within a specified distance of any given road.
}
\details{
Uses paralellization
}
