% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_mvcs.R
\name{process_mvcs}
\alias{process_mvcs}
\title{Process MVCs to add streets information}
\usage{
process_mvcs(
  mvcs_filename = here::here("../data/MVCs_Export.csv"),
  streets_filename = here::here("../data/STREETS 2/STREETS.shp"),
  streets_layer = "STREETS",
  mvc_coords = c("scene_lon", "scene_lat")
)
}
\arguments{
\item{mvcs_filename}{Source CSV file for MVCs}

\item{streets_filename}{Source shapefile for Streets}

\item{streets_layer}{layer to use from the streets shapefile}

\item{mvc_coords}{Column names in the MVCs file for lat/lon}
}
\value{
sf object fully merged
}
\description{
Takes in MVCs incident information from ePCR,
streets shapefile, and then outputs merged fileset.
}
