% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imports.R
\name{read_mvc_data}
\alias{read_mvc_data}
\title{Read MVC Incident Data}
\usage{
read_mvc_data(filename = here::here("../data/MVCs_Export.csv"))
}
\arguments{
\item{filename}{Filename of data to pull in.}
}
\value{
tibble with imported and corrected file.
}
\description{
Reads the MVC data from an ImageTrend Bulk Export,
corrects column types, creates factors where needed,
and outputs a tibble with the necessary contents.
}
